# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple medium-like blog API, writen in Go",
        "title": "Go-Blog",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/blog/{username}/posts/{post_slug}/save/{list_slug}": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Add post to user's list by providing required data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Add post to my list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blog owner's username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "post slug",
                        "name": "post_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "list slug you want to add this post to",
                        "name": "list_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
        "/lists/my": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get current user's lists by providing required data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Get current user's lists",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Create a new list for user by providing required data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Create a new list for user",
                "parameters": [
                    {
                        "description": "body required to create a new list",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ListRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
        "/lists/my/{list_slug}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get posts in my list by providing required data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Get posts in my list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "list slug you want to get",
                        "name": "list_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update current user's list information (name and description) by providing the list ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Update current user's list by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "list slug you want to edit",
                        "name": "list_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body to update",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.ListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete current user's list by providing the list ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Delete current user's list by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "list slug you want to remove",
                        "name": "list_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
        "/lists/my/{list_slug}/{post_slug}": {
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Remove post from current user's list by providing the list ID and post slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Remove post from current user's list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "list slug you want to remove post from",
                        "name": "list_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "post slug you want to delete",
                        "name": "post_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "objects"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ListRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "JWT Bearer Token. Need to Login to get the token. Usage: \"Bearer \u003cyour-token-here\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}